@@ -1,13 +1,14 @@
 import React from "react";
 import { storiesOf, addDecorator } from "@kadira/storybook";
-import { withKnobs, text, boolean, number, array, select, object } from "@kadira/storybook-addon-knobs";
+import { withKnobs, text, boolean, number, array, select } from "@kadira/storybook-addon-knobs";
 import withReadme from "storybook-readme/with-readme";
 
-import NestedListDefault from "./NestedList.stories";
 import NestedListReadme from "@appbaseio/reactivemaps-manual/docs/v1/components/NestedList.md";
 
-require ("../../node_modules/materialize-css/dist/css/materialize.min.css");
-require ("../../dist/css/style.min.css");
+import NestedListDefault from "./NestedList.stories";
+
+require("../../node_modules/materialize-css/dist/css/materialize.min.css");
+require("../../dist/css/style.min.css");
 
 function removeFirstLine(str) {
 	return str.substring(str.indexOf("\n") + 1);
@@ -20,16 +21,18 @@ storiesOf("NestedList", module)
 	)))
 	.add("With Title", withReadme(removeFirstLine(NestedListReadme), () => (
 		<NestedListDefault
-			title={text("title", "Car Category")} />
+			title={text("title", "Car Category")}
+		/>
 	)))
 	.add("Default selection", withReadme(removeFirstLine(NestedListReadme), () => (
 		<NestedListDefault
-			defaultSelected={["bmw", "x series"]} />
+			defaultSelected={["bmw", "x series"]}
+		/>
 	))).add("Playground", withReadme(removeFirstLine(NestedListReadme), () => (
 		<NestedListDefault
 			title={text("title", "NestedList: Car Filter")}
 			size={number("size", 100)}
-			sortBy={select("sortBy", {asc: "asc", desc: "desc", count: "count"}, "count")}
+			sortBy={select("sortBy", { asc: "asc", desc: "desc", count: "count" }, "count")}
 			defaultSelected={array("defaultSelected", ["bmw", "x series"])}
 			showCount={boolean("showCount", true)}
 			showSearch={boolean("showSearch", true)}
